examples/finance-agent/pyproject.toml
[project]
name = "finance_agent"
version = "0.1.0"
description = "An AI agent project built with rayai"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "rayai",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*"]
